jni::objects

Enum JValueGen

pub enum JValueGen<O> {
    Object(O),
    Byte(jbyte),
    Char(jchar),
    Short(jshort),
    Int(jint),
    Long(jlong),
    Bool(jboolean),
    Float(jfloat),
    Double(jdouble),
    Void,
}

Rusty version of the JNI C jvalue enum. Used in Java method call arguments and returns.

JValueGen is a generic type, meant to represent both owned and borrowed JNI values. The type parameter O refers to what kind of object reference the JValueGen can hold, which is either:

an owned JObject, used for values returned from a Java method call, or

a borrowed &JObject, used for parameters passed to a Java method call.

These two cases are represented by the type aliases JValueOwned and JValue, respectively.

Variants

Object(O)

Byte(jbyte)

Char(jchar)

Short(jshort)

Int(jint)

Long(jlong)

Bool(jboolean)

Float(jfloat)

Double(jdouble)

Void

Implementations

impl<O> JValueGen<O>

pub fn as_jni<'local>(&self) -> jvalue
where O: AsRef<JObject<'local>> + Debug,

Convert the enum to its jni-compatible equivalent.

pub fn to_jni<'local>(self) -> jvalue
where O: AsRef<JObject<'local>> + Debug,

๐Deprecated: Use as_jni instead.

Convert the enum to its jni-compatible equivalent.

pub fn type_name(&self) -> &'static str

Get the type name for the enum variant.

pub fn primitive_type(&self) -> Option<Primitive>

Get the primitive type for the enum variant. If itโs not a primitive (i.e. an Object), returns None.

pub fn l(self) -> Result<O>

Try to unwrap to an Object.

pub fn z(self) -> Result<bool>

Try to unwrap to a boolean.

pub fn b(self) -> Result<jbyte>

Try to unwrap to a byte.

pub fn c(self) -> Result<jchar>

Try to unwrap to a char.

pub fn d(self) -> Result<jdouble>

Try to unwrap to a double.

pub fn f(self) -> Result<jfloat>

Try to unwrap to a float.

pub fn i(self) -> Result<jint>

Try to unwrap to an int.

pub fn j(self) -> Result<jlong>

Try to unwrap to a long.

pub fn s(self) -> Result<jshort>

Try to unwrap to a short.

pub fn v(self) -> Result<()>

Try to unwrap to a void.

pub fn borrow(&self) -> JValueGen<&O>

Copies or borrows the value in this JValue.

If the value is a primitive type, it is copied. If the value is an object reference, it is borrowed.