jni::objects

Trait AsJArrayRaw

pub unsafe trait AsJArrayRaw<'local>: AsRef<JObject<'local>> {
    // Provided method
    fn as_jarray_raw(&self) -> jarray { ... }
}

Trait to access the raw jarray pointer for types that wrap an array reference

Safety

Implementing this trait will allow a type to be passed to JNIEnv::get_array_length() or other JNI APIs that only work with a valid reference to an array (or null)

Provided Methods

fn as_jarray_raw(&self) -> jarray

Returns the raw JNI pointer as a jarray