jni::objects

Type Alias JValue

pub type JValue<'local, 'obj_ref> = JValueGen<&'obj_ref JObject<'local>>;

A reference JValueGen.

This type is used for parameters passed to Java method calls. If the Java method is to be passed an object reference, it takes the form of a borrowed &JObject.

Aliased Type

pub enum JValue<'local, 'obj_ref> {
    Object(&'obj_ref JObject<'local>),
    Byte(i8),
    Char(u16),
    Short(i16),
    Int(i32),
    Long(i64),
    Bool(u8),
    Float(f32),
    Double(f64),
    Void,
}

Variants

Object(&'obj_ref JObject<'local>)

Byte(i8)

Char(u16)

Short(i16)

Int(i32)

Long(i64)

Bool(u8)

Float(f32)

Double(f64)

Void