Module jni::signature

Parser for java type signatures.

Structs

TypeSignature

A method type signature. This is the structure representation of something like (Ljava/lang/String;)Z. Used by the call_(object|static)_method functions on jnienv to ensure safety.

Enums

JavaType

Enum representing any java type in addition to method signatures.

Primitive

A primitive java type. These are the things that can be represented without an object.