Struct jni::strings::JNIStr

pub struct JNIStr { /* fields omitted */ }

Wrapper for std::ffi::CStr that also takes care of encoding between UTF-8 and Java’s Modified UTF-8.

Implementations

impl JNIStr

pub unsafe fn from_ptr<'a>(ptr: *const c_char) -> &'a JNIStr

Construct a reference to a JNIStr from a pointer. Equivalent to CStr::from_ptr.

Safety

Expects a valid pointer to a null-terminated C string and does not perform any lifetime checks for the resulting value.