jni

Struct AttachGuard

pub struct AttachGuard<'local> { /* private fields */ }

A RAII implementation of scoped guard which detaches the current thread when dropped. The attached JNIEnv can be accessed through this guard via its Deref implementation.