jni::descriptors

Trait Desc

pub unsafe trait Desc<'local, T> {
    type Output: AsRef<T>;

    // Required method
    fn lookup(self, _: &mut JNIEnv<'local>) -> Result<Self::Output>;
}

Trait for things that can be looked up through the JNI via a descriptor. This will be something like the fully-qualified class name java/lang/String or a tuple containing a class descriptor, method name, and method signature. For convenience, this is also implemented for the concrete types themselves in addition to their descriptors.

Safety

Implementations of this trait must return the correct value from the lookup method. It must not, for example, return a random JMethodID or the JClass of a class other than the one requested. Returning such an incorrect value results in undefined behavior. This requirement also applies to the returned value’s implementation of AsRef<T>.

Required Associated Types

type Output: AsRef<T>

The type that this Desc returns.

Required Methods

fn lookup(self, _: &mut JNIEnv<'local>) -> Result<Self::Output>

Look up the concrete type from the JVM.

Note that this method does not return exactly T. Instead, it returns some type that implements AsRef<T>. For this reason, it is often necessary to use turbofish syntax when calling this method:

// The value returned by `lookup` is not exactly `JClass`.
let class/*: impl AsRef<JClass> */ =
    Desc::<JClass>::lookup("java/lang/Object", env)?;

// But `&JClass` can be borrowed from it.
let class: &JClass = class.as_ref();

Warning: Many built-in implementations of this trait return AutoLocal from this method. If you then call JObject::as_raw on the returned object reference, this may result in the reference being deleted before it is used, causing undefined behavior.

For example, don’t do this:

// Undefined behavior: the `JClass` is dropped before the raw pointer
// is passed to `some_function`!
some_function(Desc::<JClass>::lookup("java/lang/Object", env)?.as_raw());

Instead, do this:

let class = Desc::<JClass>::lookup("java/lang/Object", env)?;

some_function(class.as_raw());

drop(class);

This will still work without the call to drop at the end, but calling drop ensures that the reference is not accidentally dropped earlier than it should be.