jni

Module objects

Wrappers for object pointers returned from the JVM.

Structs

AutoElements

Auto-release wrapper for a mutable pointer to the elements of a JPrimitiveArray (such as JByteArray)

AutoElementsCritical

Auto-release wrapper for a mutable pointer to the elements of a JPrimitiveArray (such as JByteArray)

AutoLocal

Auto-delete wrapper for local refs.

GlobalRef

A global JVM reference. These are “pinned” by the garbage collector and are guaranteed to not get collected until released. Thus, this is allowed to outlive the JNIEnv that it came from and can be used in other threads.

JByteBuffer

Lifetime’d representation of a jobject that is an instance of the ByteBuffer Java class. Just a JObject wrapped in a new class.

JClass

Lifetime’d representation of a jclass. Just a JObject wrapped in a new class.

JFieldID

Wrapper around jfieldID that implements Send + Sync since method IDs are valid across threads (not tied to a JNIEnv).

JList

Wrapper for JObjects that implement java/util/List. Provides methods to get, add, and remove elements.

JListIter

An iterator over the keys and values in a java.util.List. See JList::iter for more information.

JMap

Wrapper for JObjects that implement java/util/Map. Provides methods to get and set entries and a way to iterate over key/value pairs.

JMapIter

An iterator over the keys and values in a map. See JMap::iter for more information.

JMethodID

Wrapper around jmethodID that implements Send + Sync since method IDs are valid across threads (not tied to a JNIEnv).

JObject

Wrapper around [sys::jobject] that adds a lifetime to ensure that the underlying JNI pointer won’t be accessible to safe Rust code if the object reference is released.

JObjectArray

Lifetime’d representation of a jobjectArray which wraps a JObject reference

JPrimitiveArray

Lifetime’d representation of a jarray which wraps a JObject reference

JStaticFieldID

Wrapper around jfieldID that implements Send + Sync since field IDs are valid across threads (not tied to a JNIEnv).

JStaticMethodID

Wrapper around jmethodID that implements Send + Sync since method IDs are valid across threads (not tied to a JNIEnv).

JString

Lifetime’d representation of a jstring. Just a JObject wrapped in a new class.

JThrowable

Lifetime’d representation of a jthrowable. Just a JObject wrapped in a new class.

WeakRef

A weak global JVM reference. These are global in scope like GlobalRef, and may outlive the JNIEnv they came from, but are not guaranteed to not get collected until released.

Enums

JValueGen

Rusty version of the JNI C jvalue enum. Used in Java method call arguments and returns.

ReleaseMode

ReleaseMode

Traits

AsJArrayRaw

Trait to access the raw jarray pointer for types that wrap an array reference

TypeArray

A sealed trait to define type array access/release for primitive JNI types

Type Aliases

JBooleanArray

Lifetime’d representation of a crate::sys::jbooleanArray which wraps a JObject reference

JByteArray

Lifetime’d representation of a crate::sys::jbyteArray which wraps a JObject reference

JCharArray

Lifetime’d representation of a crate::sys::jcharArray which wraps a JObject reference

JDoubleArray

Lifetime’d representation of a crate::sys::jdoubleArray which wraps a JObject reference

JFloatArray

Lifetime’d representation of a crate::sys::jfloatArray which wraps a JObject reference

JIntArray

Lifetime’d representation of a crate::sys::jintArray which wraps a JObject reference

JLongArray

Lifetime’d representation of a crate::sys::jlongArray which wraps a JObject reference

JShortArray

Lifetime’d representation of a crate::sys::jshortArray which wraps a JObject reference

JValue

A reference JValueGen.

JValueOwned

An owned JValueGen.