Struct jni::objects::AutoLocal

pub struct AutoLocal<'a: 'b, 'b> { /* private fields */ }

Auto-delete wrapper for local refs.

Anything passed to a foreign method and returned from JNI methods is considered a local ref unless it is specified otherwise. These refs are automatically deleted once the foreign method exits, but it’s possible that they may reach the JVM-imposed limit before that happens.

This wrapper provides automatic local ref deletion when it goes out of scope.

NOTE: This comes with some potential safety risks. DO NOT use this to wrap something unless you’re SURE it won’t be used after this wrapper gets dropped. Otherwise, you’ll get a nasty JVM crash.

See also the JNI specification for details on referencing Java objects and some extra information.

Implementations

impl<'a, 'b> AutoLocal<'a, 'b>

pub fn new(env: &'b JNIEnv<'a>, obj: JObject<'a>) -> Self

Creates a new auto-delete wrapper for a local ref.

Once this wrapper goes out of scope, the delete_local_ref will be called on the object. While wrapped, the object can be accessed via the Deref impl.

pub fn forget(self) -> JObject<'a>

Forget the wrapper, returning the original object.

This prevents delete_local_ref from being called when the AutoLocal gets dropped. You must either remember to delete the local ref manually, or be ok with it getting deleted once the foreign method returns.

pub fn as_obj<'c>(&self) -> JObject<'c>where
'a: 'c,

Get a reference to the wrapped object

Unlike forget, this ensures the wrapper from being dropped while the returned JObject is still live.