Struct jni::objects::JList

pub struct JList<'a: 'b, 'b> { /* fields omitted */ }

Wrapper for JObjects that implement java/util/List. Provides methods to get, add, and remove elements.

Looks up the class and method ids on creation rather than for every method call.

Implementations

impl<'a: 'b, 'b> JList<'a, 'b>

pub fn from_env(env: &'b JNIEnv<'a>, obj: JObject<'a>) -> Result<JList<'a, 'b>>

Create a map from the environment and an object. This looks up the necessary class and method ids to call all of the methods on it so that exra work doesn’t need to be done on every method call.

pub fn get(&self, idx: jint) -> Result<Option<JObject<'a>>>

Look up the value for a key. Returns Some if it’s found and None if a null pointer would be returned.

pub fn add(&self, value: JObject<'a>) -> Result<()>

Append an element to the list

pub fn insert(&self, idx: jint, value: JObject<'a>) -> Result<()>

Insert an element at a specific index

pub fn remove(&self, idx: jint) -> Result<Option<JObject<'a>>>

Remove an element from the list by index

pub fn size(&self) -> Result<jint>

Get the size of the list

pub fn pop(&self) -> Result<Option<JObject<'a>>>

Pop the last element from the list

Note that this calls size() to determine the last index.

pub fn iter(&self) -> Result<JListIter<'a, 'b, '_>>

Get key/value iterator for the map. This is done by getting the EntrySet from java and iterating over it.