jni::objects

Struct JObject

pub struct JObject<'local> { /* private fields */ }

Wrapper around [sys::jobject] that adds a lifetime to ensure that the underlying JNI pointer won’t be accessible to safe Rust code if the object reference is released.

It matches C’s representation of the raw pointer, so it can be used in any of the extern function argument positions that would take a jobject.

Most other types in the objects module deref to this, as they do in the C representation.

The lifetime 'local represents the local reference frame that this reference belongs to. See the JNIEnv documentation for more information about local reference frames. If 'local is 'static, then this reference does not belong to a local reference frame, that is, it is either null or a global reference.

Note that an owned JObject is always a local reference and will never have the 'static lifetime. GlobalRef does implement AsRef<JObject<’static>>, but this only yields a borrowed &JObject<'static>, never an owned JObject<'static>.

Local references belong to a single thread and are not safe to share across threads. This type implements Send and Sync if and only if the lifetime 'local is 'static.

Implementations

impl<'local> JObject<'local>

pub unsafe fn from_raw(raw: jobject) -> Self

Creates a JObject that wraps the given raw jobject

Safety

raw may be a null pointer. If raw is not a null pointer, then:

raw must be a valid raw JNI local reference.

There must not be any other JObject representing the same local reference.

The lifetime 'local must not outlive the local reference frame that the local reference was created in.

pub fn as_raw(&self) -> jobject

Returns the raw JNI pointer.

pub fn into_raw(self) -> jobject

Unwrap to the internal jni type.

pub fn null() -> JObject<'static>

Creates a new null reference.

Null references are always valid and do not belong to a local reference frame. Therefore, the returned JObject always has the 'static lifetime.