Struct jni::objects::JObject

#[repr(transparent)]
pub struct JObject<'a> { /* private fields */ }

Wrapper around sys::jobject that adds a lifetime. This prevents it from outliving the context in which it was acquired and getting GC’d out from under us. It matches C’s representation of the raw pointer, so it can be used in any of the extern function argument positions that would take a jobject.

Most other types in the objects module deref to this, as they do in the C representation.

Implementations

impl<'a> JObject<'a>

pub unsafe fn from_raw(raw: jobject) -> Self

Creates a JObject that wraps the given raw jobject

Safety

Expects a valid pointer or null

pub fn into_raw(self) -> jobject

Unwrap to the internal jni type.

pub fn null() -> JObject<'a>

Creates a new null object