Module jni::strings

String types for going to/from java strings.

Structs

JNIStr

Wrapper for std::ffi::CStr that also takes care of encoding between UTF-8 and Java’s Modified UTF-8.

JNIString

Wrapper for std::ffi::CString that also takes care of encoding between UTF-8 and Java’s Modified UTF-8. As with CString, this implements Deref to &JNIStr.

JavaStr

Reference to a string in the JVM. Holds a pointer to the array returned by GetStringUTFChars. Calls ReleaseStringUTFChars on Drop. Can be converted to a &JNIStr with the same cost as the &CStr.from_ptr conversion.