Struct jni::strings::JavaStr

pub struct JavaStr<'a: 'b, 'b> { /* private fields */ }

Reference to a string in the JVM. Holds a pointer to the array returned by GetStringUTFChars. Calls ReleaseStringUTFChars on Drop. Can be converted to a &JNIStr with the same cost as the &CStr.from_ptr conversion.

Implementations

impl<'a: 'b, 'b> JavaStr<'a, 'b>

pub fn from_env(env: &'b JNIEnv<'a>, obj: JString<'a>) -> Result<Self>

Build a JavaStr from an object and a reference to the environment. You probably want to use JNIEnv::get_string instead.

pub fn get_raw(&self) -> *const c_char

Extract the raw C string pointer from the JavaStr. This will be encoded using the JVM internal CESU-8-style.